requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-google-genai
langgraph
msal
pydantic-settings
requests
python-dotenv
pytest
pytest-mock
httpx
jinja2
python-multipart
streamlit